add combo rom
