+++ examples/combo/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "combo"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+agb = { version = "0.12.2", path = "../../agb" }
+the-purple-night = { path = "../the-purple-night" }
+the-hat-chooses-the-wizard = { path = "../the-hat-chooses-the-wizard" }
+hyperspace-roll = { path = "../hyperspace-roll" }
+
+
+[profile.dev]
+opt-level = 2
+debug = true
+
+[profile.release]
+panic = "abort"
+lto = true
+debug = true
