+++ examples/combo/gfx/games.toml
@@ -1,0 +1,16 @@
+version = "1.0"
+
+[image.hat]
+filename = "hat.png"
+tile_size = "8x8"
+transparent_colour = "121105"
+
+[image.purple]
+filename = "purple.png"
+tile_size = "8x8"
+transparent_colour = "121105"
+
+[image.hyperspace]
+filename = "hyperspace.png"
+tile_size = "8x8"
+transparent_colour = "121105"
